Interplay between complex symmetry and Koenigs eigenfunctions S. Waleed Noor, Osmar R. Severiano
    We investigate the relationship between the complex symmetry of
composition operators Cϕf=f∘ϕ induced on the classical Hardy space
H2(𝔻) by an analytic self-map ϕ of the open unit disk 𝔻 and its
Koenigs eigenfunction. A generalization of orthogonality known as
conjugate-orthogonality will play a key role in this work. We show that
if ϕ is a Schröder map (fixes a point a∈𝔻 with 0<|ϕ′(a)|<1) and σ is
its Koenigs eigenfunction, then Cϕ is complex symmetric if and only if
(σn)n∈ℕ is complete and conjugate-orthogonal in H2(𝔻). We study the
conjugate-orth