Fix bin/lookup to do what you probably want
Use a for loop to apply dict to each argument of bin/lookup and tee the
output to a file named for the argument in the "words" directory in the
user's home directory. Create the ${HOME}/words directory if it does
not exist.
Warning: dict is not documented as a dependency of this project.