arb4j  is the Licensed Work
by Author and Licensor  Stephen Crowley ©2023 <crow@outsider.trading>

that is made available under the terms of the

Business Source License™ v1.1 ©2023

by MariaDB plc, All Rights Reserved

Definition 1. The Licensed Work is the complete set of files, including
source code, that come along with and associated with the arb4j
distribution that this license accompanies; including transformations of
the source-code (such as but not limited to the compiled.class files or
the dynamically generated .class files generated by the expression
compilation capabilities of arb4j)
Additional Use Grant
Personal, educational, and hobby use of the Licensed Work is allowed
without a commercial license. However, if the Licensed Work is used
within a company or non-profit organization for more than two weeks, a
commercial license is required. The use of the Licensed Work(arb4j) for
commercial product development of any kind, especially financial
purposes (including option pricing, futures pricing, hedging, or similar
financial calculations), regardless of whether coinciding with being a
personal, educational, or hobby use is prohibited and would be an
instance of Production Use, thus requiring a license. Exceptions to this
for the purposes of testing and evaluation must be approved in writng by
the Licensor
Terms
The Licensor(Stephen Crowley) hereby grants you the right to
•  copy,
•  modify,
•  create derivative works,
•  redistribute,
•  and make non-production use of the Licensed Work.
The Licensor may make an Additional Use Grant, above, permitting limited
production use.
Change Date
Effective on the Change Date of December 18, 2023+4 years, or the fourth
anniversary of the first publicly available distribution of a specific
version of the Licensed Work under this License, whichever comes first,
the Licensor hereby grants you rights under the terms of the Change
License, and the rights granted in the paragraph above terminate.

This software shall be licensed under the Business Source License v1.1
until 4 years after the date of December 18, 2023. On this Change Date,
the software will automatically transition to the Lesser General Public
License (LGPL).

If your use of the Licensed Work does not comply with the requirements
currently in effect as described in this License, you must purchase a
commercial license from the Licensor, its affiliated entities, or
authorized resellers, or you must refrain from using the Licensed Work.
All copies of the original and modified Licensed Work, and derivative
works of the Licensed Work, are subject to this License. This License
applies separately for each version of the Licensed Work and the Change
Date may vary for each version of the Licensed Work released by
Licensor.
You must conspicuously display this License on each original or modified
copy of the Licensed Work. If you receive the Licensed Work in original
or modified form from a third party, the terms and conditions set forth
in this License apply to your use of that work.
Any use of the Licensed Work in violation of this License will
automatically terminate your rights under this License for the current
and all other versions of the Licensed Work.
This License does not grant you any right in any trademark or logo of
Licensor or its affiliates (provided that you may use a trademark or
logo of Licensor as expressly required by this License).TO THE EXTENT
PERMITTED BY APPLICABLE LAW, THE LICENSED WORK IS PROVIDED ON AN “AS IS”
BASIS. LICENSOR HEREBY DISCLAIMS ALL WARRANTIES AND CONDITIONS, EXPRESS
OR IMPLIED, INCLUDING (WITHOUT LIMITATION) WARRANTIES OF
MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
TITLE. MariaDB hereby grants you permission to use this License's text
to license your works, and to refer to it using the trademark “Business
Source License”, as long as you comply with the Covenants of Licensor
below.
Covenants of Licensor
In consideration of the right to use this License's text and the
“Business Source License” name and trademark, Licensor covenants to
MariaDB, and to all other recipients of the licensed work to be provided
by Licensor:
To specify as the Change License the GPL Version 2.0 or any later
version, or a license that is compatible with GPL Version 2.0 or a later
version, where “compatible” means that software provided under the
Change License can be included in a program with software provided under
GPL Version 2.0 or a later version. Licensor may specify additional
Change Licenses without limitation.To either:
a) specify an additional grant of rights to use that does not impose any
additional restriction on the right granted in this License, as the
Additional Use Grant;

or

b) insert the text “None” to specify a Change Date. Not to modify this
License in any other way.

Notice
The Business Source License(this document, or the “License”) is not an
Open Source license. However, the Licensed Work will eventually be made
available under an Open Source License, as stated in this License.
For more information on the use of the Business Source License for
MariaDB products, please visit the MariaDB Business Source License FAQ.
For more information on the use of the Business Source License
generally, please visit the Adopting and Developing Business Source
License FAQ.