arb.expressions.nodes.unary.FunctionCall<D, R, F extends Function<D, R>>
FunctionCall is a Node in the Expression that represents a call to
either a builtin or a contextual function. A built-in function is one
that is a member of the type of the domain or range of the function,
such as Real.tanh(int, Real).

A contextual function is one that is shared thru a mutual Context by
constructing one and passing it to the Function.express(Class, Class,
String, Context) function or one of the other variations of the express
function that accepts a Context argument.

For the function to be callable from another function via a shared
Context either the Function.express(Class, Class, String, String,
Context) or Function.express(Class, Class, String, String, Context,
boolean) methods that accepts another String argument specifying the
name of the function should be used.

 arb4j is made available under the terms of the Business Source License™
v1.1
 ©2024 which can be found in the root directory of this project in a
file
 named License.pdf, License.txt, or License.tm which are the pdf, text,
and
 TeXmacs formatted versions of the same document respectively.