Stationary Isotropic CorrelStationary Isotropic Correlation Functions
Abstract
Examples of Stationary Isotropic Correlation Functions in 1, 2 and 3
dimensions are recalled from Yaglom'87[1]

Example 1. [1, Example 6]Consider the stationary correlation function
(2.101) given by the equation
B(τ)=C*exp (-α*τ)*cos (ω_0*τ)  (1)
where α>0, ω_0>0, τ>0.
It is clear that if α is small enough, this function does not satisfy
even the first inequality (4.1.13), i.e. it does not belong to the class
D_0 of isotropic correlation function in the space ℝ^n. By evaluating,
with the aid of (4.1.18) or (2.102) and (4.1.32), the two-dimensional
spectral density f(k), which corresponds to the correlation function
B(τ), and by studying the conditions for the nonnegativity of this f(k),
one can show that the function (2.101) is a possible correlation
function of the isotropic field in the space ℝ^n if, and only if,
α>sqrt(ω_0^2+λ_n^2). It is still easier to obtain conditions
guaranteeing that B(τ) belongs to D_n. Indeed, it was noted on pp.
124-125 that for a (one-dimensional) spectral density corresponding to
the correlation function (2.101) is monotonically nonincreasing on the
positive half-axis if, and only if, α≥sqrt(3)*ω_0. It follows that the
function (2.101) is a three-dimensional isotropic correlation function
in the space ℝ^3 if, and only if, α≥sqrt(3)*ω_0.
Example 2. [1, Example 7]Formula (4.105) shows that the function
B(τ)=(*J_((n-2)/2)*(α*τ))/(C*(α*τ)^((n-2)/2))  (2)
where C>0, α>0, is an n-dimensional isotropic correlation function. The
spectral distribution function Φ(k) corresponding to this correlation
function is clearly of the form
Φ(k)={
0,                                 for k<α
(C/2(α/2)^(n/2))/(Γ(n/2)*k^(n+1))* for k>α
.
i.e., the n-dimensional spectral density is given here by the relation
f(k)=Φ'(k)(α/2)^(n/2)*2*π^(-n/2)*k^(n-1)=(C/((2*π)^(n/2))*α^(n/2)*)/(Γ*(n/2+1))*k^(n-1)*δ*(k-α)
(3)
where δ(x) is the Dirac δ-function. Using (4.120), we easily obtain an
equation of the form (2.138), with ν=(n-2)/2, for the corresponding
one-dimensional spectral density f_1(k). Replacing (n-2)/2 in the
right-hand side of (4.1.39) by an arbitrary real number ν>-1/2, we
arrive at a more general stationary correlation function (2.137). The
inequality (4.112) clearly implies that the function (2.137) does not
belong to the class D_n (i.e. cannot be an isotropic correlation
function in the n-dimensional space ℝ^n) if ν<(n-2)/2. On the other
hand, it is not hard to show that if ν>(n-2)/2 the function (2.137)
belongs to the class D_n . In fact, it is well known that for ν>ν_∗=-1
the following identity holds:
int_0^∞k^μ*J_μ*(k*x)*J_μ*(k*y)*k*
d*k=(2^(μ-1)*Γ*(ν+μ))/(x^μ*Γ(μ)*y^(ν+1))*δ*(x-y) ∀μ>-1/2  (4)
(see, e.g., Erdelyi et al., 1953, Vol. 2, Eq. 7.14.2(34) or Gradshteyn
and Ryzhik, 1980, Eq. 6.575(1)). According to (4.114) it follows from
this that, for ν>(n-2)/2, the n-dimensional spectral density f(k), which
corresponds to the correlation function (2.137), has the form
f(k)={
C*(α^2-k^2)ν-n/2 for k<α
0                for k>α
.  (5)
where C is a normalization constant. The function (4.139) is everywhere
nonnegative; hence the function (2.137), where ν>(n-2)/2, belongs to
D_n.
Some more examples of isotropic correlation functions can also be found
in the available literature, but we will not pursue this topic any
further here.
Bibliography
[1]  A.M. Yaglom. Correlation Theory of Stationary and Related Random
Functions: Volume I: Basic Results. Applied Probability. Springer New
York, 1987.ation Functions