The Bessel Polynomial Moments
1. Introduction
Since 1949 when the Bessel polynomials appeared, a hunt has been on to
find a weight function ψ, defined on the real axis, with respect to
which the polynomials would be orthogonal. Although at least two devices
have been found which formally seem to fill the role of orthogonality
generation, none was expressible in terms of integration on the real
axis.
In a paper by Krall [1] it was shown that the Bessel polynomials
y_n(x,a,b) are orthogonal with respect to the distribution ψ given by
⟨f,ψ⟩=lim_(ε→0)
1/π*int_α^βf(x)*Im{(-b/(x+i*ε))*_1F_1(a,-b/(x+i*ε))}*d*x∀α<0,β>0  (1)
The key to showing that the above formula is valid is through the
moments associated with the Bessel polynomials
μ_n=((-b)^(n+1))/((a)_n)    (2)
where (a)_n=a*(a+1)*…*(a+n-1)∀n=0,1,… are the rising factorials, which
were discovered by examining the differential equation satisfied by the
Bessel polynomials. It is well known that, given the moments μ_n, there
exists a function of a complex variable f(z), called the Cauchy
representation, given by
f(z)=sum_(n=0)^∞(μ_n)/(z^(n+1))  (3)
when the support of the weight function is compact, and in general
satisfying
lim_(z→∞) z^(k+1)*[f(z)+sum_(n=0)^k(μ_n)/(z^(n+1))]=-μ_k ∀δ<arg
z<π-δ∃δ∈(0,π/2)  (4)
For the Bessel polynomials it is easy to see that
f(z)=sum_(n=0)^∞((-b/z)^(n+1))/((a)_n)=(-b)/z/ _1F_1(
[1],[a]
,-b/z)  (5)
Bibliography
[1]  A.M. Krall. The besel polynomial moment problem. Acta Mathematica
Academiae Scientiarum Hungarica, 38:105–107, 1981.