Kellog's Method of Constructing Characteristic Values and Characteristic Elements
The following method of constructing the characteristic values different
from 0 of a completely continuous symmetric transformation A≠0 is due to
KELLOGG and communicated in [1, 96] by Riesz&Nagy.
Let f_0 be an element such that A*f_0≠0, but otherwise arbitrary. Set
f_n=A^n*f_0 (n=1,2,…)  (1)
The relation
(U*f_n,f_n) =(A*f_(n-1),f_n)
            =(f_(n-1),A*f_n)
            =(f_(n-1),f_(n+1))
(n=1,2,…)  (2)
makes it clear that if f_n≠0, then also f_(n+1)≠0. Since f_0≠0 and
f_1≠0, we have ∀n:f_n≠0. Denoting the ratio (‖f_(n+1)‖)/(‖f_n‖) by r_n
and setting ‖f_0‖=v_0 we have
f_n=(v_0*r_1*…*r_(n-1))*g_n  (3)
where g_n is an element of 𝔖 of norm 1. We have the equations
A*g_n=r_(n+1)*g_(n+1)+A^2*g_n=r_(n+1)*r_(n+2)*g_(n+2)+…  (4)
It follows from the first equation that r_(n+1)≤‖A‖. On the other hand,
(15) implies that
r_n=((g_(n-1),f_(n+1)))/(‖f_n‖)≤r_n+1  (5)
Consequently the sequence {r_n} converges to some positive quantity r.
Making use of equations (16), we obtain
‖g_n-g_(n+2)‖^2 =2-2*(g_n,g_(n+2))
                =2-(g_(n+1)+A^2*g_n,g_(n+1)+A^2*g_n)
                =2-2(((A*g_n,A*g_n))/(r_(n+1)*r_(n+2)))
                =2-2((r_(n+1)^2)/(r_(n+1)*r_(n+2)))
                =2*(1-(r_(n+1))/(r_(n+2)))→0
(6)
hence
g_n-g_(n+2)=g_n-(A^2*g_n)/(r_(n+1)*r_(n+2))*→0*as n→∞  (7)
Since the sequence of elements
h_n=(g_n)/(r_(n+1)+r_(n+2))*  (8)
is bounded, it contains a subsequence for which A^2*h_n converges to
some element φ of 𝔖. The corresponding subsequence of {g_n} converges,
by (1), to the same element φ, and that of the sequence
{g_n-(A^2*g_n)/(r_(n+1)+r_(n+2))*}  (9)
to
φ-(A^2*φ)/r*  (10)
therefore we have
φ-(A^2*φ)/r*=φ  (11)
A^2*φ=r*φ  (12)
Since φ is the limit of the sequence {g_n} of elements of norm 1, φ has
norm 1 also. Therefore φ is a characteristic element of A^2. We obtain a
characteristic element of A from it in the following manner. We set
A*φ=r*ψ  (13)
A*χ=-r*χ  (14)
then
A*ψ=r*φ  (15)
A*χ=-r*χ  (16)
Since φ+χ=2*ψ≠0, the elements φ and χ cannot be zero simultaneously,
hence at least one of the values r and -r is a characteristic value of
A.
Note 1. More profound study of the sequence {g_n}, as well as an
application of the same method to certain non-completely continuous
symmetric transformations, is found in the Notes [1], [2] of R. Wavre.
The same method can be applied with success to non-symmetric completely
continuous linear transformations which can be made symmetric if
multiplied on the left by a suitable positive symmetric transformation
(symmetrizable transformations).
Cf. Zaanen [1]. Integral equations with "symmetrizable" kernels have
been studied by several authors, starting with Hilbert (Note 5); see the
article in the Encyclopädie by Hellinger and Toeplitz [2] (pp.
1536–1543) and, among recent contributions, see Reid [1], where
bibliographical references are also found.
Bibliography
[1]  Frigyes Riesz and Béla Sz.-Nagy. Functional Analysis. Dover Books
on Mathematics. Dover Publications, unabridged republication of 1955
edition, 2012.