The Gaussian Cosmic Microwave Background Conjecture
by Stephen Crowley

March 14, 2024μ

The BOOMERanG (Balloon Observations Of Millimetric Extragalactic
Radiation ANd Geophysics) experiment, conducted in 1998, significantly
improved the estimate of the curvature of the Universe by providing
precise measurements of the Cosmic Microwave Background (CMB)
anisotropies. The pivotal findings from this experiment, particularly
regarding the Universe's geometry as flat, were announced in April 2000.
These results implied that the total energy density of the universe is
very close to the critical density, with Ω_(t*o*t*a*l)≈1, where
Ω_(t*o*t*a*l) is the total energy density of the Universe divided by the
critical density. This was a significant advancement, supporting
inflationary theory predictions and leading to a concordance model of
cosmology.
The temperature fluctuations in the Cosmic Microwave Background (CMB)
across the sky follow a Gaussian random field, indicating that these
fluctuations are statistically isotropic and homogeneous with a zero
mean and fully characterized by their variance.Conjecture 1